Siirrä koodit palveluvayla servletiin
