+++ src/main/resources/reference.conf
@@ -118,6 +118,7 @@ features = {
 }
 
 suomi-fi-user-oid = "1.2.246.562.24.99999988889"
+hsl-user-oid = "1.2.246.562.24.99999988899"
 
 // These mock values need to be overridden. Otherwise mock data will be used.
 virta.serviceUrl="mock"
