+++ src/main/scala/fi/oph/koski/koskiuser/RequiresSuomiFiOrHsl.scala
@@ -2,7 +2,7 @@ package fi.oph.koski.koskiuser
 
 import fi.oph.koski.http.KoskiErrorCategory
 
-trait RequiresSuomiFi extends KoskiSpecificAuthenticationSupport {
+trait RequiresSuomiFiOrHsl extends KoskiSpecificAuthenticationSupport {
   implicit def koskiSession: KoskiSpecificSession = koskiSessionOption.get
 
   before() {
@@ -10,8 +10,8 @@ trait RequiresSuomiFi extends KoskiSpecificAuthenticationSupport {
       case Left(status) if status.statusCode == 401 =>
         haltWithStatus(status)
       case _ =>
-        if (!koskiSessionOption.exists(_.hasSuomiFiAccess)) {
-          haltWithStatus(KoskiErrorCategory.forbidden.vainDVV())
+        if (!koskiSessionOption.exists(_.hasSuomiFiAccess) && !koskiSessionOption.exists(_.hasHSLAccess)) {
+          haltWithStatus(KoskiErrorCategory.forbidden.vainPalveluvayla())
         }
     }
   }
